Fix bug 2004850 : do not try to use libnvidia-ml.so.
Since we are loading the library at runtime, we should only ever use
libnvidia-ml.so.1.
The ".so" library is only there for compiling user programs ; at runtime
we should always use the ".so.1" because we have embedded the "1" API and
if the version changes, the code would likely be broken.